Refactor service entrypoints.
