+++ app/bin/tools/backfill_history.dart
@@ -10,9 +10,9 @@ import 'package:gcloud/db.dart';
 
 import 'package:pub_dartlang_org/account/backend.dart';
 import 'package:pub_dartlang_org/frontend/models.dart';
-import 'package:pub_dartlang_org/frontend/service_utils.dart';
 import 'package:pub_dartlang_org/history/backend.dart';
 import 'package:pub_dartlang_org/history/models.dart';
+import 'package:pub_dartlang_org/service/entrypoint/tools.dart';
 
 final _argParser = ArgParser()
   ..addOption('package', abbr: 'p', help: 'The package to backfill.');
