+++ app/bin/tools/check_integrity.dart
@@ -11,7 +11,7 @@ import 'package:pool/pool.dart';
 
 import 'package:pub_dartlang_org/account/models.dart';
 import 'package:pub_dartlang_org/frontend/models.dart';
-import 'package:pub_dartlang_org/frontend/service_utils.dart';
+import 'package:pub_dartlang_org/service/entrypoint/tools.dart';
 
 final _argParser = ArgParser()
   ..addOption('concurrency',
