+++ app/bin/tools/clear_package_properties.dart
@@ -10,7 +10,7 @@ import 'package:gcloud/db.dart';
 import 'package:pool/pool.dart';
 
 import 'package:pub_dartlang_org/frontend/models.dart';
-import 'package:pub_dartlang_org/frontend/service_utils.dart';
+import 'package:pub_dartlang_org/service/entrypoint/tools.dart';
 
 final _argParser = ArgParser()
   ..addOption('concurrency',
