+++ app/bin/tools/flag.dart
@@ -8,11 +8,11 @@ import 'dart:io';
 import 'package:args/args.dart';
 import 'package:gcloud/db.dart';
 
+import 'package:pub_dartlang_org/frontend/backend.dart';
 import 'package:pub_dartlang_org/frontend/models.dart';
-import 'package:pub_dartlang_org/frontend/service_utils.dart';
+import 'package:pub_dartlang_org/service/entrypoint/tools.dart';
 import 'package:pub_dartlang_org/shared/analyzer_client.dart';
 import 'package:pub_dartlang_org/shared/redis_cache.dart' show cache;
-import 'package:pub_dartlang_org/frontend/backend.dart';
 
 final _argParser = ArgParser(allowTrailingOptions: true)
   ..addFlag('help',
