+++ app/bin/tools/package_stats.dart
@@ -15,7 +15,7 @@ import 'package:gcloud/db.dart';
 
 import 'package:pub_dartlang_org/frontend/models.dart';
 import 'package:pub_dartlang_org/frontend/model_properties.dart';
-import 'package:pub_dartlang_org/frontend/service_utils.dart';
+import 'package:pub_dartlang_org/service/entrypoint/tools.dart';
 
 Future main(List<String> args) async {
   final ArgParser parser = ArgParser()
