+++ app/bin/tools/pubspec_stats.dart
@@ -8,7 +8,7 @@ import 'package:appengine/appengine.dart';
 import 'package:gcloud/db.dart';
 
 import 'package:pub_dartlang_org/frontend/models.dart';
-import 'package:pub_dartlang_org/frontend/service_utils.dart';
+import 'package:pub_dartlang_org/service/entrypoint/tools.dart';
 
 Future main(List<String> args) async {
   useLoggingPackageAdaptor();
