--- app/lib/frontend/service_utils.dart
@@ -1,59 +1,0 @@
-// Copyright (c) 2014, the Dart project authors.  Please see the AUTHORS file
-// for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
-
-library pub_dartlang_org.server_common;
-
-import 'dart:async';
-
-import 'package:http/http.dart' as http;
-import 'package:shelf/shelf.dart' as shelf;
-
-import '../account/backend.dart';
-import '../shared/configuration.dart';
-import '../shared/services.dart';
-
-import 'upload_signer_service.dart';
-
-/// Looks at [request] and if the 'Authorization' header was set tries to get
-/// the user email address and registers it.
-Future registerLoggedInUserIfPossible(shelf.Request request) async {
-  final authorization = request.headers['authorization'];
-  if (authorization != null) {
-    final parts = authorization.split(' ');
-    if (parts.length == 2 && parts.first.trim().toLowerCase() == 'bearer') {
-      final accessToken = parts.last.trim();
-
-      final user =
-          await accountBackend.authenticateWithAccessToken(accessToken);
-      if (user != null) {
-        registerAuthenticatedUser(user);
-      }
-    }
-  }
-}
-
-Future<String> obtainServiceAccountEmail() async {
-  final http.Response response = await http.get(
-      'http://metadata/computeMetadata/'
-      'v1/instance/service-accounts/default/email',
-      headers: const {'Metadata-Flavor': 'Google'});
-  return response.body.trim();
-}
-
-/// Helper for utilities in bin/tools to setup a minimal AppEngine environment,
-/// calling [fn] to run inside it. It registers only the most frequently used
-/// services (at the moment only `frontend/backend.dart`).
-///
-/// Connection parameters are inferred from the GCLOUD_PROJECT and the GCLOUD_KEY
-/// environment variables.
-Future withProdServices(Future fn()) {
-  return withServices(() {
-    if (!envConfig.hasCredentials) {
-      throw Exception('Missing GCLOUD_* environments for package:appengine');
-    }
-    registerUploadSigner(
-        ServiceAccountBasedUploadSigner(activeConfiguration.credentials));
-    return fn();
-  });
-}
