+++ app/lib/service/isolate.dart
@@ -3,20 +3,18 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'dart:async';
-import 'dart:io';
 import 'dart:isolate';
 
 import 'package:appengine/appengine.dart';
 import 'package:logging/logging.dart';
 import 'package:meta/meta.dart';
-import 'package:pana/pana.dart' show runProc;
 import 'package:stack_trace/stack_trace.dart';
 
-import 'configuration.dart';
-import 'scheduler_stats.dart';
+import '../shared/configuration.dart';
+import '../shared/scheduler_stats.dart';
+import '../shared/utils.dart' show trackEventLoopLatency;
+
 import 'services.dart';
-import 'utils.dart' show trackEventLoopLatency;
-import 'versions.dart';
 
 class FrontendEntryMessage {
   final int frontendIndex;
@@ -185,36 +183,6 @@ Future startIsolates({
 void setupServiceIsolate() {
   useLoggingPackageAdaptor();
   trackEventLoopLatency();
-}
-
-Future initFlutterSdk(Logger logger) async {
-  if (envConfig.flutterSdkDir == null) {
-    logger.warning('FLUTTER_SDK is not set, assuming flutter is in PATH.');
-  } else {
-    // If the script exists, it is very likely that we are inside the appengine.
-    // In local development environment the setup should happen only once, and
-    // running the setup script multiple times should be safe (no-op if
-    // FLUTTER_SDK directory exists).
-    if (FileSystemEntity.isFileSync('/project/app/script/setup-flutter.sh')) {
-      final sw = Stopwatch()..start();
-      logger.info('Setting up flutter checkout. This may take some time.');
-      final ProcessResult result = await runProc(
-          '/project/app/script/setup-flutter.sh', ['v$flutterVersion'],
-          timeout: const Duration(minutes: 5));
-      if (result.exitCode != 0) {
-        throw Exception(
-            'Failed to checkout flutter (exited with ${result.exitCode})\n'
-            'stdout: ${result.stdout}\nstderr: ${result.stderr}');
-      }
-      final flutterBin = File('${envConfig.flutterSdkDir}/bin/flutter');
-      if (!(await flutterBin.exists())) {
-        throw Exception(
-            'Flutter binary is missing after running setup-flutter.sh');
-      }
-      sw.stop();
-      logger.info('Flutter checkout completed in ${sw.elapsed}.');
-    }
-  }
 }
 
 void _wrapper(List fnAndMessage) {
