+++ pkg/fake_pub_server/lib/fake_pub_server.dart
@@ -23,7 +23,7 @@ import 'package:pub_dartlang_org/shared/configuration.dart';
 import 'package:pub_dartlang_org/shared/handler_helpers.dart';
 import 'package:pub_dartlang_org/shared/redis_cache.dart';
 import 'package:pub_dartlang_org/shared/search_client.dart';
-import 'package:pub_dartlang_org/shared/services.dart';
+import 'package:pub_dartlang_org/service/services.dart';
 
 final _logger = Logger('fake_server');
